requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
openai
rapidfuzz
redis
django-redis
mysqlclient
requests
beautifulsoup4
lxml
